updated color tokens for light/dark mode and momentum component use improvements
