+++ src/components/WebexWalkin/WebexWalkin.ts
@@ -22,6 +22,10 @@ export namespace WebexWalkin {
     @property({ attribute: "brand-name" }) brandName: string | null = null;
     @property({ attribute: "agent-id" }) agentId: string | null = null;
 
+    @property({ type: Number }) seconds = 180; // seconds
+    progressValue: number | null = null;
+    intervalID: any;
+
     fullScreen = false;
     webex: any;
     public isWebexMeetingConnected = false;
@@ -30,6 +34,16 @@ export namespace WebexWalkin {
 
     profile: any | null = null;
     isAudioOnly = false;
+
+    connectedCallback() {
+      super.connectedCallback();
+      this.startTimer();
+    }
+
+    disconnectedCallback() {
+      super.disconnectedCallback();
+      clearInterval(this.intervalID);
+    }
 
     updated(changedProperties: PropertyValues) {
       super.updated(changedProperties);
@@ -42,6 +56,10 @@ export namespace WebexWalkin {
           }
         }, 500);
       }
+
+      if (changedProperties.has("seconds")) {
+        this.startTimer();
+      }
     }
 
     getSpinner() {
@@ -49,7 +67,7 @@ export namespace WebexWalkin {
         <div class="spinner">
           ${this.isAuthDenied
             ? html`
-                <span>Unable to connect to server</span>
+                <span class="error-message">Unable to connect to server</span>
               `
             : html`
                 <md-spinner size="32"></md-spinner>
@@ -58,10 +76,47 @@ export namespace WebexWalkin {
       `;
     }
 
+    public startTimer() {
+      this.progressValue = 100;
+      if (this.intervalID !== undefined) {
+        clearInterval(this.intervalID);
+      }
+
+      this.intervalID = setInterval(() => {
+        this.progressValue = (this.progressValue as number) - (10 * 100) / (this.seconds * 1000);
+
+        this.requestUpdate();
+
+        if (this.progressValue <= 0) {
+          this.progressValue = 0;
+          clearInterval(this.intervalID);
+          const event = new CustomEvent("timed-out", {
+            composed: true,
+            bubbles: true
+          });
+
+          this.dispatchEvent(event);
+        }
+      }, 10);
+    }
+
+    renderProgressBar = () => {
+      return this.progressValue
+        ? html`
+            <md-progress-bar
+              .type=${"determinate"}
+              .value=${this.progressValue}
+              .displayFormat=${"none"}
+              .dynamic=${true}
+            ></md-progress-bar>
+          `
+        : nothing;
+    };
+
     getWelcomeScreen() {
       return html`
         <div class="stretch stack-two-column">
-          <cjaas-timer></cjaas-timer>
+          ${this.renderProgressBar()}
           <div class="banner">
             <div class="profile-details">
               <div class="display-name">${this.profile.nickName}</div>
@@ -135,7 +190,7 @@ export namespace WebexWalkin {
       return html`
         <div class="call-container connecting">
           <div class=${this.fullScreen ? "self-video-container fullscreen" : "self-video-container"}>
-            <div class=${this.isAudioOnly ? "calling-info dark" : "calling-info"}>
+            <div class="calling-info">
               Calling ${this.profile.nickName}
             </div>
             ${this.isAudioOnly
@@ -150,9 +205,9 @@ export namespace WebexWalkin {
             <video id="remote-view-video" autoplay></video>
           </div>
           <md-button
-            .color=${"color-none"}
+            color="color-none"
             id="resize"
-            .circle=${true}
+            circle
             .size=${this.fullScreen ? "68" : "44"}
             @click=${() => {
               this.resizeModal();
@@ -163,27 +218,27 @@ export namespace WebexWalkin {
           <div class=${this.fullScreen ? "call-controls-answered fullscreen" : "call-controls-answered"}>
             <md-button
               id="screen-share"
-              .circle=${true}
+              circle
               .size=${this.fullScreen ? "68" : "44"}
               title="Share Screen"
-              .color=${"blue"}
+              variant="primary"
             >
-              <md-icon .name=${"share-screen_24"}></md-icon>
+              <md-icon name="share-screen_24"></md-icon>
             </md-button>
             <md-button
               id="stop-share"
-              .circle=${true}
+              circle
               .size=${this.fullScreen ? "68" : "44"}
               title="Stop Sharing"
-              .color=${"blue"}
+              variant="primary"
             >
-              <md-icon .name=${"stop-content-share_20"}></md-icon>
+              <md-icon name="stop-content-share_20"></md-icon>
             </md-button>
             <md-button
               id="hang-up"
-              .circle=${true}
+              circle
               .size=${this.fullScreen ? "68" : "44"}
-              .color=${"red"}
+              variant="red"
               title="Hang Up"
               @click="${() => {
                 this.showMeetingControls = false;
@@ -191,7 +246,7 @@ export namespace WebexWalkin {
                 this.requestUpdate();
               }}"
             >
-              <md-icon .name=${"cancel_24"}></md-icon>
+              <md-icon name="cancel_24"></md-icon>
             </md-button>
           </div>
         </div>
@@ -200,14 +255,14 @@ export namespace WebexWalkin {
 
     getControls() {
       return html`
-        <md-button .color=${"green"} .circle=${true} @click=${() => this.callAgent()} .size=${"44"}
-          ><md-icon .name=${"camera_24"}></md-icon>
+        <md-button variant="green" circle @click=${() => this.callAgent()} size="44"
+          ><md-icon slot="icon" name="camera_24"></md-icon>
         </md-button>
-        <md-button .color=${"green"} .circle=${true} .size=${"44"} @click=${() => this.callAgent(true)}
-          ><md-icon .name=${"handset_24"}></md-icon>
+        <md-button variant="green" circle size="44" @click=${() => this.callAgent(true)}
+          ><md-icon slot="icon" name="handset_24"></md-icon>
         </md-button>
-        <md-button .color=${"red"} .circle=${true} .size=${"44"} @click=${() => this.dispatchCloseEvent()}
-          ><md-icon .name=${"cancel_24"}></md-icon>
+        <md-button variant="red" circle size="44" @click=${() => this.dispatchCloseEvent()}
+          ><md-icon slot="icon" name="cancel_24"></md-icon>
         </md-button>
       `;
     }
