+++ src/components/WebexWalkin/scss/webexWalkin.scss
@@ -58,6 +58,7 @@
 }
 
 .invite {
+  background-color: var(--md-primary-bg-color, $md-white);
   display: flex;
   align-items: flex-end;
   flex-direction: column;
@@ -68,7 +69,7 @@
 }
 
 .message {
-  background: #e9e9e9;
+  background-color: var(--md-tertiary-bg-color, $md-gray-10);
   border-radius: 12px;
   margin: 12px;
   padding: 8px;
@@ -82,7 +83,7 @@
   height: 100%;
   position: relative;
   box-sizing: border-box;
-  background: white;
+  background-color: var(--md-primary-bg-color, $md-white);
   font-family: "CiscoSansTT Regular", "Helvetica Neue", Helvetica, Arial, sans-serif;
 }
 
@@ -149,9 +150,10 @@ div {
 }
 
 .spinner {
-  display: flex;
   align-items: center;
+  display: flex;
   justify-content: center;
+  margin-top: 2rem;
 }
 .stretch {
   height: 100%;
@@ -209,6 +211,7 @@ div {
 }
 
 .connecting .self-video-container {
+  background-color: var(--md-primary-bg-color, $md-white);
   height: 100%;
   width: 100%;
   right: 0;
@@ -226,7 +229,7 @@ div {
   top: 12px;
   width: 100%;
   text-align: center;
-  color: white;
+  color: var(--md-primary-text-color, $md-gray-100);
   font-size: 16px;
   z-index: 10;
   text-decoration: 2px 2px 4px black;
